split owner info files in two for other system than linux
