+++ receiver/filelogreceiver/README.md
@@ -33,8 +33,8 @@ Tails and parses logs from files.
 | `include_file_path`                 | `false`                              | Whether to add the file path as the attribute `log.file.path`.                                                                                                                                                                                                  |
 | `include_file_name_resolved`        | `false`                              | Whether to add the file name after symlinks resolution as the attribute `log.file.name_resolved`.                                                                                                                                                               |
 | `include_file_path_resolved`        | `false`                              | Whether to add the file path after symlinks resolution as the attribute `log.file.path_resolved`.                                                                                                                                                               |
-| `include_file_owner_name`           | `false`                              | Whether to add the file owner name as the attribute `log.file.owner.name`  |
-| `include_file_group_name`           | `false`                              | Whether to add the file group name as the attribute `log.file.group.name`  |
+| `include_file_owner_name`           | `false`                              | Whether to add the file owner name as the attribute `log.file.owner.name`. Only supported for linux or solaris.  |
+| `include_file_group_name`           | `false`                              | Whether to add the file group name as the attribute `log.file.group.name`. Only supported for linux or solaris.  |
 | `poll_interval`                     | 200ms                                | The [duration](#time-parameters) between filesystem polls.                                                                                                                                                                                                      |
 | `fingerprint_size`                  | `1kb`                                | The number of bytes with which to identify a file. The first bytes in the file are used as the fingerprint. Decreasing this value at any point will cause existing fingerprints to forgotten, meaning that all files will be read from the beginning (one time) |
 | `max_log_size`                      | `1MiB`                               | The maximum size of a log entry to read. A log entry will be truncated if it is larger than `max_log_size`. Protects against reading large amounts of data into memory.                                                                                         |
